Correction of modifications
