+++ bootstrapers/showcase/mappers/blogsData.js
@@ -1,9 +1,9 @@
-import BlogSection from '../../../lib/organisms/components/portfolios/portfolio4';
+import CharityBlogSection from '../../../lib/organisms/components/portfolios/portfolio4';
 
 const resolveComponents = () => {
   return {
     403389: {
-      render: props => <BlogSection {...props} />,
+      render: props => <CharityBlogSection {...props} />,
       name: 'Blog Section Data Section',
       category: 'Section',
     },
