+++ bootstrapers/showcase/mappers/branch.js
@@ -1,9 +1,9 @@
-import BranchSection from '../../../lib/organisms/components/portfolios/portfolio5';
+import  CharityBranchSection from '../../../lib/organisms/components/portfolios/portfolio5';
 
 const resolveComponents = () => {
   return {
     403117: {
-      render: props => <BranchSection {...props} />,
+      render: props => <CharityBranchSection {...props} />,
       name: 'Branch Section',
       category: 'Section',
     },
