+++ bootstrapers/showcase/mappers/clients.js
@@ -1,15 +1,15 @@
-import ClientsSection from '../../../lib/organisms/components/clients/client1';
-import ClientBlock from '../../../lib/organisms/components/sliders/slider5';
+import CharityClientsSection from '../../../lib/organisms/components/clients/client1';
+import CharityClientBlock from '../../../lib/organisms/components/sliders/slider5';
 
 const resolveComponents = () => {
   return {
     403300: {
-      render: props => <ClientsSection {...props} />,
+      render: props => <CharityClientsSection {...props} />,
       name: 'Portfolio Clients',
       category: 'Clients',
     },
     403134: {
-      render: props => <ClientBlock {...props} />,
+      render: props => <CharityClientBlock {...props} />,
       name: 'Portfolio Clients',
       category: 'Clients',
     },
