+++ bootstrapers/showcase/mappers/donationData.js
@@ -1,9 +1,9 @@
-import DonateSection from '../../../lib/organisms/components/callToActions/callToAction6';
+import CharityDonateSection from '../../../lib/organisms/components/callToActions/callToAction6';
 
 const resolveComponents = () => {
   return {
     403126: {
-      render: props => <DonateSection {...props} />,
+      render: props => <CharityDonateSection {...props} />,
       name: 'Donate Section Data Section',
       category: 'Section',
     },
