+++ bootstrapers/showcase/mappers/fundraiserData.js
@@ -1,9 +1,9 @@
-import FundraiserSection from '../../../lib/organisms/components/callToActions/callToAction7';
+import CharityFundraiserSection from '../../../lib/organisms/components/callToActions/callToAction7';
 
 const resolveComponents = () => {
   return {
     403387: {
-      render: props => <FundraiserSection {...props} />,
+      render: props => <CharityFundraiserSection {...props} />,
       name: 'Fundraiser Section Data Section',
       category: 'Section',
     },
