+++ bootstrapers/showcase/mappers/mapData.js
@@ -1,9 +1,9 @@
-import MapSection from '../../../lib/organisms/components/portfolios/portfolio6';
+import CharityMapSection from '../../../lib/organisms/components/imageDetails/imageDetail1';
 
 const resolveComponents = () => {
   return {
     403386: {
-      render: props => <MapSection {...props} />,
+      render: props => <CharityMapSection {...props} />,
       name: 'Map Section Data Section',
       category: 'Section',
     },
