+++ bootstrapers/showcase/mappers/workData.js
@@ -1,9 +1,9 @@
-import WorkSection from '../../../lib/organisms/components/featuresLists/featureList19';
+import CharityWorkSection from '../../../lib/organisms/components/featuresLists/featureList19';
 
 const resolveComponents = () => {
   return {
     403119: {
-      render: props => <WorkSection {...props} />,
+      render: props => <CharityWorkSection {...props} />,
       name: 'Work Data Section',
       category: 'Section',
     },
