+++ lib/organisms/components/banners/banner8/drawer/style.ts
@@ -1,12 +1,50 @@
 import styled from 'styled-components';
 import { themeGet } from 'styled-system';
-import { Button } from '../../../navbars/navbar8/style';
 
 const InnerWrapper = styled.div`
   min-height: 100vh;
   padding: 30px;
 `;
 
+export const Button = styled.button`
+  border: 0;
+  padding: 0;
+  min-width: 150px;
+  min-height: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 50px;
+  font-size: 14px;
+  font-weight: 600;
+  cursor: pointer;
+  color: ${themeGet('colors.heading', '#060F1E')};
+  background-color: ${themeGet('colors.primary', '#FCF22B')};
+  transition: all 0.3s ease;
+  @media only screen and (max-width: 1440px) {
+    font-size: 15px;
+  }
+  @media only screen and (max-width: 1360px) {
+    font-size: 14px;
+  }
+
+  &:hover {
+    background-color: ${themeGet('colors.primaryHover', '#ECF22F')};
+  }
+
+  img {
+    margin-left: 11px;
+    @media only screen and (max-width: 1440px) {
+      margin-left: 10px;
+    }
+  }
+
+  &:hover,
+  &:focus {
+    content: 0;
+    box-shadow: none;
+  }
+`;
 export const SpreadButton = styled(Button)`
   display: none;
   @media only screen and (max-width: 991px) {
