+++ lib/organisms/components/imageDetails/imageDetail1/style.ts
@@ -1,0 +1,67 @@
+import styled from 'styled-components';
+
+const SectionWrapper = styled.section`
+  width: 100%;
+  padding: 81px 0;
+  @media only screen and (max-width: 1440px) {
+    padding: 70px 0;
+  }
+  @media only screen and (max-width: 1360px) {
+    padding: 45px 0 50px;
+  }
+  @media only screen and (max-width: 991px) {
+    padding: 40px 0;
+  }
+`;
+
+export const SectionHeader = styled.header`
+  text-align: center;
+  margin-bottom: 90px;
+  @media only screen and (max-width: 1440px) {
+    margin-bottom: 70px;
+  }
+  @media only screen and (max-width: 991px) {
+    margin-bottom: 54px;
+  }
+
+  h2 {
+    font-size: 50px;
+    font-weight: 300;
+    margin-bottom: 12px;
+    @media only screen and (max-width: 1440px) {
+      font-size: 36px;
+    }
+    @media only screen and (max-width: 1360px) {
+      font-size: 34px;
+    }
+    @media only screen and (max-width: 991px) {
+      font-size: 30px;
+      margin-bottom: 10px;
+    }
+    @media only screen and (max-width: 767px) {
+      font-size: 28px;
+    }
+  }
+
+  p {
+    font-size: 18px;
+    line-height: 40px;
+    margin-bottom: 0;
+    @media only screen and (max-width: 1440px) {
+      font-size: 16px;
+      line-height: 32px;
+    }
+    @media only screen and (max-width: 991px) {
+      font-size: 15px;
+      line-height: 28px;
+    }
+  }
+`;
+
+export const ImageWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
+
+export default SectionWrapper;
