+++ tsconfig.json
@@ -5,7 +5,10 @@
     "allowJs": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-    "types": ["node", "jest"],
+    "types": [
+      "node",
+      "jest"
+    ],
     "strict": false,
     "forceConsistentCasingInFileNames": true,
     "noEmit": true,
@@ -18,6 +21,9 @@
     "**/*.tsx",
     "**/*.js",
     "**/*.jsx",
-    "next-env.d.ts", "lib/organisms/components/footers/footer7/index.tsx", ],
-  "exclude": ["node_modules"]
+    "next-env.d.ts",
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
